+++ glue/app/qt/application.py
@@ -19,6 +19,7 @@ from glue.icons.qt import get_icon
 from glue.utils.qt import get_qapp
 from glue.app.qt.actions import action
 from glue.dialogs.data_wizard.qt import data_wizard
+from glue.dialogs.link_editor.qt import LinkEditor
 from glue.app.qt.edit_subset_mode_toolbar import EditSubsetModeToolBar
 from glue.app.qt.mdi_area import GlueMdiArea, GlueMdiSubWindow
 from glue.app.qt.layer_tree_widget import PlotAction, LayerTreeWidget
@@ -33,7 +34,7 @@ from glue.utils.qt import (pick_class, GlueTabBar,
 from glue.app.qt.feedback import submit_bug_report, submit_feedback
 from glue.app.qt.plugin_manager import QtPluginManager
 from glue.app.qt.versions import show_glue_info
-from glue.app.qt.terminal import glue_terminal, IPythonTerminalError
+from glue.app.qt.terminal import glue_terminal
 
 
 __all__ = ['GlueApplication']
@@ -73,6 +74,11 @@ def _fix_ipython_pylab():
     except UsageError:
         pass
 
+    # Make sure we disable interactive mode (where figures get redrawn for
+    # every single Matplotlib command)
+    import matplotlib
+    matplotlib.interactive(False)
+
 
 class GlueLogger(QtWidgets.QWidget):
 
@@ -98,7 +104,13 @@ class GlueLogger(QtWidgets.QWidget):
         report = QtWidgets.QPushButton("Send Bug Report")
         report.clicked.connect(nonpartial(self._send_report))
 
-        self.stderr = sys.stderr
+        if isinstance(sys.stderr, GlueLogger):
+            if isinstance(sys.stderr._stderr_original, GlueLogger):
+                raise Exception('Too many nested GlueLoggers')
+            self._stderr_original = sys.stderr._stderr_original
+        else:
+            self._stderr_original = sys.stderr
+
         sys.stderr = self
 
         l = QtWidgets.QVBoxLayout()
@@ -117,7 +129,7 @@ class GlueLogger(QtWidgets.QWidget):
 
     def _set_console_button(self, attention):
         if attention:
-            self.button_console.setStyleSheet('background: red;')
+            self.button_console.setStyleSheet('color: red; text-decoration: underline;')
         else:
             self.button_console.setStyleSheet(self.button_stylesheet)
 
@@ -125,7 +137,7 @@ class GlueLogger(QtWidgets.QWidget):
         """
         Interface for sys.excepthook
         """
-        self.stderr.write(message)
+        self._stderr_original.write(message)
         self._text.moveCursor(QtGui.QTextCursor.End)
         self._text.insertPlainText(message)
         self._set_console_button(attention=True)
@@ -164,6 +176,10 @@ class GlueLogger(QtWidgets.QWidget):
         """
         if event.key() == Qt.Key_Escape:
             self.hide()
+
+    def closeEvent(self, event):
+        if sys.stderr is self:
+            sys.stderr = self._stderr_original
 
 
 class GlueApplication(Application, QtWidgets.QMainWindow):
@@ -226,20 +242,85 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
         self._ui.data_layers.setLayout(self._vb)
         self._layer_widget = lw
 
-        self._ui.layout_top.setSpacing(0)
+        # Data toolbar
 
-        # log window + status light
-        self._log = GlueLogger(button_console=self._ui.button_console)
+        self._data_toolbar = QtWidgets.QToolBar()
+
+        self._data_toolbar.setIconSize(QtCore.QSize(16, 16))
+
+        self._button_open_data = QtWidgets.QToolButton()
+        self._button_open_data.setText("Open Data")
+        self._button_open_data.setIcon(get_icon('glue_open'))
+        self._button_open_data.setToolButtonStyle(Qt.ToolButtonTextBesideIcon)
+        self._button_open_data.clicked.connect(nonpartial(self._choose_load_data))
+
+        self._data_toolbar.addWidget(self._button_open_data)
+
+        self._button_link_data = QtWidgets.QToolButton()
+        self._button_link_data.setText("Link Data")
+        self._button_link_data.setIcon(get_icon('glue_link'))
+        self._button_link_data.setToolButtonStyle(Qt.ToolButtonTextBesideIcon)
+        self._button_link_data.clicked.connect(self._set_up_links)
+
+        self._data_toolbar.addWidget(self._button_link_data)
+
+        self._button_ipython = QtWidgets.QToolButton()
+        self._button_ipython.setCheckable(True)
+        self._button_ipython.setText("IPython Terminal")
+        self._button_ipython.setIcon(get_icon('IPythonConsole'))
+        self._button_ipython.setToolButtonStyle(Qt.ToolButtonTextBesideIcon)
+        self._button_ipython.clicked.connect(self._toggle_terminal)
+
+        self._data_toolbar.addWidget(self._button_ipython)
+
+        spacer = QtWidgets.QWidget()
+        spacer.setMinimumSize(20, 10)
+        spacer.setSizePolicy(QtWidgets.QSizePolicy.Fixed,
+                             QtWidgets.QSizePolicy.Preferred)
+
+        self._data_toolbar.addWidget(spacer)
+
+        self.addToolBar(self._data_toolbar)
+
+        # Selection mode toolbar
+
+        tbar = EditSubsetModeToolBar()
+        self._mode_toolbar = tbar
+
+        self.addToolBar(self._mode_toolbar)
+
+        # Error console toolbar
+
+        self._console_toolbar = QtWidgets.QToolBar()
+
+        spacer = QtWidgets.QWidget()
+        spacer.setSizePolicy(QtWidgets.QSizePolicy.Expanding,
+                             QtWidgets.QSizePolicy.Preferred)
+
+        self._console_toolbar.addWidget(spacer)
+
+        self._button_console = QtWidgets.QToolButton()
+        self._button_console.setText("View Error Console")
+        self._button_console.setToolButtonStyle(Qt.ToolButtonTextOnly)
+
+        self._console_toolbar.addWidget(self._button_console)
+
+        self.addToolBar(self._console_toolbar)
+
+        self._log = GlueLogger(button_console=self._button_console)
         self._log.window().setWindowTitle("Console Log")
         self._log.resize(550, 550)
         self._log.hide()
+
+    def _set_up_links(self, event):
+        LinkEditor.update_links(self.data_collection)
 
     def _tweak_geometry(self, maximized=True):
         """Maximize window by default."""
         if maximized:
             self.setWindowState(Qt.WindowMaximized)
-        self._ui.main_splitter.setStretchFactor(0, 0)
-        self._ui.main_splitter.setStretchFactor(1, 1)
+        self._ui.main_splitter.setStretchFactor(0, 0.1)
+        self._ui.main_splitter.setStretchFactor(1, 0.9)
         self._ui.data_plot_splitter.setStretchFactor(0, 0.25)
         self._ui.data_plot_splitter.setStretchFactor(1, 0.5)
         self._ui.data_plot_splitter.setStretchFactor(2, 0.25)
@@ -475,24 +556,6 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
         mbar.addMenu(menu)
 
         menu = QtWidgets.QMenu(mbar)
-        menu.setTitle("&Toolbars")
-        tbar = EditSubsetModeToolBar()
-        tbar.setStyle(QtWidgets.QStyleFactory.create("windows"));
-        self._mode_toolbar = tbar
-        self._ui.layout_top.insertWidget(0, tbar)
-        a = QtWidgets.QAction("Selection Mode &Toolbar", menu)
-        a.setCheckable(True)
-        a.toggled.connect(tbar.setVisible)
-        try:
-            tbar.visibilityChanged.connect(a.setChecked)
-        except AttributeError:  # Qt < 4.7. QtCore.Signal not supported
-            pass
-
-        menu.addAction(a)
-        menu.addActions(tbar.actions())
-        mbar.addMenu(menu)
-
-        menu = QtWidgets.QMenu(mbar)
         menu.setTitle("&Plugins")
         menu.addAction(self._actions['plugin_manager'])
         menu.addSeparator()
@@ -723,12 +786,12 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
             self.close()
         return ga
 
-    def _reset_session(self, show=True):
+    def _reset_session(self, show=True, warn=True):
         """
         Reset session to clean state.
         """
 
-        if not os.environ.get('GLUE_TESTING'):
+        if not os.environ.get('GLUE_TESTING') and warn:
             buttons = QtWidgets.QMessageBox.Ok | QtWidgets.QMessageBox.Cancel
             dialog = QtWidgets.QMessageBox.warning(
                 self, "Confirm Close",
@@ -738,8 +801,15 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
             if not dialog == QtWidgets.QMessageBox.Ok:
                 return
 
+        # Make sure the closeEvent gets executed to close the GlueLogger
+        self._log.close()
+        self.app.processEvents()
+
         ga = GlueApplication()
-        ga.show()
+        ga.start(block=False)
+
+        # We need to close this after we open the next application otherwise
+        # Qt will quit since there are no actively open windows.
         self.close()
 
         return ga
@@ -763,14 +833,15 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
         """
         ga = Application.restore_session(path)
         if show:
-            ga.show()
+            ga.start(block=False)
         return ga
 
-    def has_terminal(self):
+    def has_terminal(self, create_if_not=True):
         """
         Returns True if the IPython terminal is present.
         """
-        self._create_terminal()  # ensure terminal is setup
+        if self._terminal is None and create_if_not:
+            self._create_terminal()
         return self._terminal is not None
 
     def _create_terminal(self):
@@ -778,48 +849,16 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
         if self._terminal is not None:  # already set up
             return
 
-        if hasattr(self, '_terminal_exception'):  # already failed to set up
-            return
+        widget = glue_terminal(data_collection=self._data,
+                               dc=self._data,
+                               hub=self._hub,
+                               session=self.session,
+                               application=self,
+                               **vars(env))
 
-        self._terminal_button = QtWidgets.QToolButton(self._ui)
-        self._terminal_button.setToolTip("Toggle IPython Prompt")
-        i = get_icon('IPythonConsole')
-        self._terminal_button.setIcon(i)
-        self._terminal_button.setIconSize(QtCore.QSize(25, 25))
-
-        self._layer_widget.ui.button_row.addWidget(self._terminal_button)
-
-        try:
-            widget = glue_terminal(data_collection=self._data,
-                                   dc=self._data,
-                                   hub=self._hub,
-                                   session=self.session,
-                                   application=self,
-                                   **vars(env))
-        except IPythonTerminalError as exc:
-            self._terminal = None
-            self._terminal_button.setEnabled(False)
-            self._terminal_button.setToolTip(str(exc))
-            self._terminal_exception = str(exc)
-            return
-
-        self._terminal_button.clicked.connect(self._toggle_terminal)
         self._terminal = self.add_widget(widget)
+        self._terminal.closed.connect(self._on_terminal_close)
         self._hide_terminal()
-
-    def _setup_terminal_error_dialog(self, exception):
-        """ Reassign the terminal toggle button to show dialog on error"""
-        title = "Terminal unavailable"
-        msg = ("Glue encountered an error trying to start the Terminal"
-               "\nReason:\n%s" % exception)
-
-        def show_msg():
-            mb = QtWidgets.QMessageBox(QtWidgets.QMessageBox.Critical,
-                             title, msg)
-            mb.setDetailedText(self._terminal_exception)
-            mb.exec_()
-
-        self._terminal_button.clicked.connect(show_msg)
 
     def _toggle_terminal(self):
         if self._terminal.isVisible():
@@ -833,6 +872,12 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
                 warnings.warn("An unexpected error occurred while "
                               "trying to show the terminal")
 
+    def _on_terminal_close(self):
+        if self._button_ipython.isChecked():
+            self._button_ipython.blockSignals(True)
+            self._button_ipython.setChecked(False)
+            self._button_ipython.blockSignals(False)
+
     def _hide_terminal(self):
         self._terminal.hide()
 
@@ -840,7 +885,7 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
         self._terminal.show()
         self._terminal.widget().show()
 
-    def start(self, size=None, position=None):
+    def start(self, size=None, position=None, block=True):
         """
         Show the GUI and start the application.
 
@@ -867,19 +912,10 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
         # XXX find out a better place for this
         _fix_ipython_pylab()
 
-        return self.app.exec_()
+        if block:
+            return self.app.exec_()
 
     exec_ = start
-
-    def keyPressEvent(self, event):
-        """Hold down modifier keys to temporarily set edit mode"""
-        mod = event.modifiers()
-        if mod == Qt.ShiftModifier:
-            self._mode_toolbar.set_mode('or')
-
-    def keyReleaseEvent(self, event):
-        """Unset any temporary edit mode"""
-        self._mode_toolbar.unset_mode()
 
     def dragEnterEvent(self, event):
         if event.mimeData().hasUrls():
@@ -907,6 +943,7 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
 
     def closeEvent(self, event):
         """Emit a message to hub before closing."""
+        self._log.close()
         self._hub.broadcast(ApplicationClosedMessage(None))
         event.accept()